requirements.txt
aiohttp==3.11.11
APScheduler==3.10.4
fastapi==0.115.6
pydantic==2.10.4
python-dotenv==1.0.1
selenium==4.27.1
uvicorn==0.34.0
web3==6.3.0
apscheduler
pandas>=1.5.0
numpy>=1.21.0
dune-client==1.3.0